Move the common TCP proxy code to stream_proxy.go
This will allow us to share with the Unix sockets implementation.
